+++ backends/aws-lambda/host/Cargo.toml
@@ -23,7 +23,6 @@ lazy_static = "1.4.0"
 once_cell = "1.4.0"
 clap = "3.0.0-beta.1"
 reqwest = { version = "0.10.4", features = ["blocking"] }
-awsio = { version = "0.1.2", package = "assemblylift-awslambda-iomod", path = "../iomod" }
 assemblylift_core = { version = "0.1.1", package = "assemblylift-core", path = "../../../core" }
 assemblylift_core_event = { version = "0.1.1", package = "assemblylift-core-event", path = "../../../core/event" }
 assemblylift_core_event_common = { version = "0.1.0", package = "assemblylift-core-event-common", path = "../../../core/event/common" }
