--- backends/aws-lambda/iomod/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "assemblylift-awslambda-iomod"
-version = "0.1.2"
-description = "AssemblyLift AWS Lambda IOmod host library"
-authors = ["Akkoro and the AssemblyLift contributors <assemblylift@akkoro.io>"]
-edition = "2018"
-license-file = "../../../LICENSE.md"
-repository = "https://github.com/akkoro/assemblylift"
-readme = "README.md"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-serde_json = "1.0.53"
-futures = "0.3.4"
-lazy_static = "1.4.0"
-paste = "0.1.12"
-crossbeam-utils = "0.7.2"
-rusoto_core = { version = "0.43.0" }
-rusoto_dynamodb = { version = "0.43.0", features = ["serialize_structs", "deserialize_structs"] }
-wasmer_runtime = { version = "0.1.1", package = "wasmer-runtime-asml-fork" }
-wasmer_runtime_core = { version = "0.1.1", package = "wasmer-runtime-core-asml-fork" }
-assemblylift_core = { version = "0.1.1", package = "assemblylift-core", path = "../../../core"}
-assemblylift_core_event = { version = "0.1.1", package = "assemblylift-core-event", path = "../../../core/event" }
-assemblylift_core_event_common = { version = "0.1.0", package = "assemblylift-core-event-common", path = "../../../core/event/common" }
