--- backends/aws-lambda/iomod/README.md
@@ -1,4 +1,0 @@
-assemblylift-awslambda-iomod
----------------------------
-
-AssemblyLift AWS Lambda IOmod host library
