+++ backends/aws-lambda/iomod/plugins/dynamodb/Cargo.toml
@@ -16,6 +16,8 @@ rusoto_core = { version = "0.43.0" }
 rusoto_dynamodb = { version = "0.43.0", features = ["serialize_structs", "deserialize_structs"] }
 
 assemblylift_core = { version = "0.1.1", package = "assemblylift-core", path = "../../../../../core"}
+assemblylift_core_event = { version = "0.1.1", package = "assemblylift-core-event", path = "../../../../../core/event" }
+assemblylift_core_iomod = { version = "0.1.0", package = "assemblylift-core-iomod", path = "../../../../../core/iomod" }
 wasmer_runtime = { version = "0.1.1", package = "wasmer-runtime-asml-fork" }
 wasmer_runtime_core = { version = "0.1.1", package = "wasmer-runtime-core-asml-fork" }
