+++ core/event/Cargo.toml
@@ -17,4 +17,8 @@ lazy_static = "1.4.0"
 indexmap = "1.3.2"
 crossbeam-utils = "0.7.2"
 serde = { version = "1.0.106", features = ["derive"]}
+
+wasmer_runtime_core = { version = "0.1.1", package = "wasmer-runtime-core-asml-fork" }
+
+assemblylift-core = { version = "0.1.1", path = "../../core" }
 assemblylift-core-event-common = { version = "0.1.0", path = "../../core/event/common" }
