+++ core/event/src/lib.rs
@@ -1,4 +1,41 @@
 #[macro_use]
 extern crate lazy_static;
 
-pub mod threader;
+use std::future::Future;
+
+use crossbeam_utils::atomic::AtomicCell;
+
+use wasmer_runtime_core::vm;
+
+use assemblylift_core_event_common::constants::EVENT_BUFFER_SIZE_BYTES;
+use assemblylift_core::{IoModulePlugin, WasmBufferPtr};
+use assemblylift_core::threader::Threader;
+
+#[inline(always)]
+pub fn spawn_event(
+    plugin_decl: &IoModulePlugin,
+    ctx: &mut vm::Ctx,
+    mem: WasmBufferPtr,
+    future: impl Future<Output=Vec<u8>> + 'static + Send,
+) -> i32 {
+    let threader: *mut Threader = ctx.data.cast();
+    if threader.is_null() {
+        panic!("Threader instance is NULL in spawn_event")
+    }
+
+    let threader_ref = unsafe { threader.as_mut().unwrap() };
+
+    let event_id = threader_ref.next_event_id().unwrap();
+    println!("DEBUG: event_id={}", event_id);
+
+    let wasm_instance_memory = ctx.memory(0);
+    let memory_writer: &[AtomicCell<u8>] =
+        match mem.deref(wasm_instance_memory, 0, EVENT_BUFFER_SIZE_BYTES as u32) {
+            Some(memory) => memory,
+            None => panic!("could not dereference WASM guest memory in spawn_event"),
+        };
+
+    threader_ref.spawn_with_event_id(plugin_decl, memory_writer.as_ptr(), future, event_id);
+
+    event_id as i32
+}
