+++ core/iomod/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "assemblylift-core-iomod"
+version = "0.1.0"
+authors = ["xlem <xlem@akkoro.io>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tokio = "0.2.21"
+crossbeam-utils = "0.7.2"
+once_cell = "1.4.0"
+lazy_static = "1.4.0"
+serde = "1.0.106"
+paste = "0.1.12"
+libloading = "0.6.2"
+wasmer_runtime = { version = "0.1.1", package = "wasmer-runtime-asml-fork" }
+wasmer_runtime_core = { version = "0.1.1", package = "wasmer-runtime-core-asml-fork" }
+assemblylift_core_event = { version = "0.1.1", package = "assemblylift-core-event", path = "../../core/event" }
+assemblylift_core_event_common = { version = "0.1.0", package = "assemblylift-core-event-common", path = "../../core/event/common" }
+
+[build-dependencies]
+rustc_version = "0.2.3"
