+++ core/iomod/src/plugin.rs
@@ -1,0 +1,32 @@
+use std::ffi::OsStr;
+use std::io;
+use std::rc::Rc;
+
+use libloading::Library;
+use once_cell::sync::Lazy;
+use tokio::runtime::Runtime;
+
+use crate::macros::{CORE_VERSION, RUSTC_VERSION};
+use std::sync::Arc;
+
+pub unsafe fn load<P: AsRef<OsStr>>(
+    mut registry: &mut ModuleRegistry,
+    library_path: P,
+) -> io::Result<IoModulePlugin> {
+    let library = Rc::new(Library::new(library_path).unwrap());
+
+    let decl = library
+        .get::<*mut IoModulePlugin>(b"__ASML_IOMOD_PLUGIN_DECL\0")
+        .unwrap()
+        .read();
+
+    if decl.rustc_version != RUSTC_VERSION || decl.asml_core_version != CORE_VERSION {
+        return Err(io::Error::new(io::ErrorKind::Other, "Version mismatch"));
+    }
+
+    println!("TRACE: loaded IOmod {}", decl.name);
+
+    (decl.register)(&mut registry);
+
+    Ok(decl)
+}
